Pre-launching of JVMs works now.
Reopen stdin/stdout/stderr in java now instead of using bash
redirection, because opening the FIFO blocks and prevents java from
launching until we connect.

I can get more speed by loading the main class before reopening the
streams, but clojure.main creates *in*/*out*/*err* when it is loaded, so
that doesn't work. I'm going to wrap the streams with a class that
delays opening the FIFO to fix this.