Passing environment to jvm now works - fixes #1
Because of the fact that the env command on OS X does not have a -0
option to delimit vars with \0 instead of newline, we cannot use
that. So instead we call declare and filter out the function definitions
and array declarations. You'll notice we do this in a bash subprocess.
This is so we can tell which env vars are exported. Unfortunately, this
isn't perfect because it seems that the bash process creates some env
variables. I'm filtering out the obvious ones for now, but hopefully
there is a better long-term solution.